refactor : Simplify the platform API + fix grid cursor centering.
Replace the dedicated grid/cursor API calls with generic drawing calls.
This was previously avoided in order to improve performance, but the
current approach adds too much maintenance overhead. Instead, we now
use caching tricks in the platform code. The exception to this is hint
logic, which still requires platform specific rendering hacks to be
performant. This should also make additional features easier to
implement.