Override default config entry
The default config entry should be overrided by the keys defined in the config file. 
In somesituation, the "button" was overrided with "j k ,", in "grid" mode although "grid_keys" was re-config to "l u n h" but "j" and "k" is still behave as "grid_keys" instead of "button" key.