Fix input processing (#171)
Currently in order to minimize conflicts each input config option is assigned
a type corresponding to the mode in which it is intended to be used. When
matching is performed, occluding bindings are discovered by looking at bindings
of the same type. Unfortunately this does not account for key definitons which
straddle multiple modes (E.g "buttons").

This patch introduces an input whitelisting mechanism which confines the match
set to the set of keys explicitly specified at the beginning of the mode.

As before, successive definitions within the config file may occlude each
other, with the most recently defined one taking precedence. This minimizes
conflicts by allowing for overlapping bindings which belong to different modes
and provides more consistent WYSIWYG config behaviour.