linux: avoid spurious warnings from make
If the artifacts to clean don't exist, rm(1) will emit a warning
which may worry the user.  But there is nothing to worry about, so
silence the warning with the -f flag.

Similarly, if the bin/ directory already exists, mkdir(1) will emit
an unnecessary warning, so add the -p flag.